backend/requirements.txt
# requirements.txt - Dependencies
Flask
mysql-connector-python
dotenv
